+++ lyricsbot/domains/lyricsondemand/__init__.py
@@ -1,12 +1,12 @@
 """
 Init.
 """
-from lyricsbot.domain.lyricsondemand.utils import (
+from lyricsbot.domains.lyricsondemand.utils import (
     remove_punctuation_symbols,
     suitable_url_parameters
 )
 
-from lyricsbot.domain.lyricsondemand.lyricsondemand import (
+from lyricsbot.domains.lyricsondemand.lyricsondemand import (
     format_request_data_url,
     parse_lyrics,
 )
