+++ lyricsbot/domains/songlyrics/__init__.py
@@ -1,12 +1,12 @@
 """
 Init.
 """
-from lyricsbot.domain.songlyrics.utils import (
+from lyricsbot.domains.songlyrics.utils import (
     remove_punctuation_symbols,
     suitable_url_parameters
 )
 
-from lyricsbot.domain.songlyrics.songlyrics import (
+from lyricsbot.domains.songlyrics.songlyrics import (
     format_request_data_url,
     parse_lyrics,
 )
